RSA Backend/RiskAssessment/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
pycodestyle==2.9.1
python-business-logic==0.2.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
